module: do not encode symbol namespace in ksymtab entries
Polluting the ksymtab entries with the symbol namespace confuses
(userland) tools - such as kmod's depmod - that depend on the format of
these entries in the binary and in System.map.

Rather than encoding the namespace in the ksymtab entry, deduct the
symbol namespaces from the __kstrtabns_<symbol> entries and their
corresponding values in __ksymtab_strings.

This refactoring also clarifies the namespace handling within modpost.
The previous interface was rather error prone and unintuitive.

Co-developed-by: Jessica Yu <jeyu@kernel.org>
Suggested-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Cc: Martijn Coenen <maco@android.com>
Cc: Shaun Ruffell <sruffell@sruffell.net>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Fixes: 8651ec01daed ("module: add support for symbol namespaces.")
Signed-off-by: Matthias Maennich <maennich@google.com>